requirements.txt
titanoboa @ git+https://github.com/vyperlang/titanoboa@c9aa6c6f31661880157de2f518f5014eab597862
boa-solidity==0.1.1
vyper==0.3.10